requirements.txt
virtualenv
django
paho-mqtt
